+++ src/storage/ct_old.h
@@ -234,7 +234,7 @@ namespace MEDDLY {
           Discard an entry and recycle the memory it used.
             @param  h   Handle of the entry.
       */
-      void discardAndRecycle(unsigned long h);
+      void discardAndRecycle(size_t h);
 
       /**
           Display an entry.
@@ -250,7 +250,7 @@ namespace MEDDLY {
             @param  s         Stream to write to.
             @param  h         Handle of the entry.
       */
-      inline void showEntry(output &s, unsigned long h) const 
+      inline void showEntry(output &s, size_t h) const 
       {
 #ifdef INTEGRATED_MEMMAN
         showEntry(s, entries+h);
@@ -667,7 +667,7 @@ void MEDDLY::ct_old<MONOLITHIC, CHAINED>::addEntry(entry_key* key, const entry_r
     }
   }
 
-  unsigned long newsize = tableSize * 2;
+  size_t newsize = tableSize * 2;
   if (newsize > maxSize) newsize = maxSize;
 
   if (CHAINED) {
@@ -1230,7 +1230,7 @@ MEDDLY::node_address MEDDLY::ct_old<MONOLITHIC, CHAINED>
 
 template <bool MONOLITHIC, bool CHAINED>
 void MEDDLY::ct_old<MONOLITHIC, CHAINED>
-::discardAndRecycle(unsigned long h)
+::discardAndRecycle(size_t h)
 {
   const int SHIFT = (MONOLITHIC ? 1 : 0) + (CHAINED ? 1 : 0);
